ref: change folder structure to feature-based
