+++ frontend/app/composables/useTheme.ts
@@ -1,49 +1,37 @@
-const THEME_COOKIE_KEY = 'user-theme-preference';
+import { ref, computed, onMounted } from 'vue';
+
 type Theme = 'light' | 'dark';
 
-export function useTheme() {
-  const themeCookie = useCookie<Theme | null>(THEME_COOKIE_KEY, {
-    maxAge: 60 * 60 * 24 * 365,
-    sameSite: 'lax',
-    path: '/',
-  });
+const currentTheme = ref<Theme>('light');
 
-  const currentTheme = useState<Theme>('app-theme', () => {
-    return themeCookie.value === 'dark' ? 'dark' : 'light';
-  });
+export const useTheme = () => {
+  const nextTheme = computed<Theme>(() => (currentTheme.value === 'light' ? 'dark' : 'light'));
 
-  // Roll forward expiry by creating a fresh useCookie instance
-  if (themeCookie.value) {
-    const currentValue = themeCookie.value;
-    // Force a new Set-Cookie by reassigning with fresh options
-    themeCookie.value = null;
-    nextTick(() => {
-      themeCookie.value = currentValue;
-    });
-  }
-
-  useHead({
-    htmlAttrs: {
-      'data-bs-theme': () => currentTheme.value,
-    },
-  });
-
-  watch(currentTheme, (newTheme) => {
-    themeCookie.value = newTheme;
+  const setTheme = (theme: Theme) => {
+    currentTheme.value = theme;
     if (import.meta.client) {
-      document.documentElement.setAttribute('data-bs-theme', newTheme);
+      document.documentElement.setAttribute('data-bs-theme', theme);
+      localStorage.setItem('theme', theme);
+    }
+  };
+
+  const toggleTheme = () => {
+    setTheme(nextTheme.value);
+  };
+
+  onMounted(() => {
+    if (import.meta.client) {
+      const savedTheme = localStorage.getItem('theme') as Theme | null;
+      const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+      const initialTheme = savedTheme || (prefersDark ? 'dark' : 'light');
+      setTheme(initialTheme);
     }
   });
 
-  const toggleTheme = () => {
-    currentTheme.value = currentTheme.value === 'light' ? 'dark' : 'light';
-  };
-
-  const nextTheme = computed(() => (currentTheme.value === 'light' ? 'dark' : 'light'));
-
   return {
     currentTheme,
-    toggleTheme,
     nextTheme,
+    toggleTheme,
+    setTheme,
   };
-}
+};
