--- frontend/app/pages/auth/login.vue
@@ -1,3 +1,0 @@
-<template>
-  <login-form />
-</template>
