+++ frontend/app/pages/auth/login/index.vue
@@ -1,0 +1,7 @@
+<template>
+    <login-form />
+</template>
+
+<script setup lang="ts">
+import LoginForm from './(components)/LoginForm.vue'
+</script>
