+++ frontend/app/pages/event-detail/(components)/sections/EventAbout.vue
