+++ frontend/app/pages/event-detail/(components)/sections/EventAds.vue
