+++ frontend/app/pages/event-detail/(components)/sections/EventHero.vue
