+++ frontend/app/pages/event-detail/(components)/sections/EventNav.vue
