+++ frontend/app/pages/event-detail/(components)/sections/EventOrganizer.vue
