+++ frontend/app/pages/event-detail/(components)/sections/EventRelated.vue
