+++ frontend/app/pages/event-detail/(components)/sections/EventSchedule.vue
