+++ frontend/app/pages/event-detail/(components)/sections/EventTickets.vue
