+++ frontend/app/pages/event-detail/index.vue
@@ -1,0 +1,7 @@
+<template>
+    <event-detail />
+</template>
+
+<script setup lang="ts">
+import EventDetail from './(components)/EventDetail.vue'
+</script>
