+++ frontend/app/pages/event-detail/types/event-detail.d.ts
