+++ frontend/app/pages/event-display/composables/use-date-range.ts
