+++ frontend/app/pages/event-display/composables/use-event-filter.ts
