+++ frontend/app/pages/event-display/data/filters.ts
