--- frontend/app/pages/event-display/display.vue
@@ -1,3 +1,0 @@
-<template>
-    <event-display/>
-</template>
