+++ frontend/app/pages/event-display/index.vue
@@ -1,0 +1,7 @@
+<template>
+    <event-display />
+</template>
+
+<script setup lang="ts">
+import EventDisplay from './(components)/EventDisplay.vue'
+</script>
