+++ frontend/app/pages/event-display/types/event-display.d.ts
