--- frontend/app/pages/event/event-detail.vue
@@ -1,3 +1,0 @@
-<template>
-  <event-detail />
-</template>
