--- frontend/app/pages/home-page/home.vue
@@ -1,3 +1,0 @@
-<template>
-  <home-page/>
-</template>
