+++ frontend/app/pages/home/(components)/EventCard.vue
