+++ frontend/app/pages/home/(components)/EventSection.vue
