+++ frontend/app/pages/home/(components)/FeaturedCarousel.vue
