+++ frontend/app/pages/home/(components)/PlaceCard.vue
