+++ frontend/app/pages/home/(components)/SpecialEventCard.vue
