+++ frontend/app/pages/home/(components)/TrendingCard.vue
