+++ frontend/app/pages/home/composables/use-carousel.ts
