+++ frontend/app/pages/home/index.vue
@@ -1,0 +1,7 @@
+<template>
+    <home-page />
+</template>
+
+<script setup lang="ts">
+import HomePage from './(components)/HomePage.vue'
+</script>
